I am now better setup for multi-week modules and testing
