+++ go.mod
@@ -1,3 +1,3 @@
-module github.com/treejamie/hackerank90
+module github.com/treejamie/hackerrank-90days
 
 go 1.24.0
