+++ test/go.mod
@@ -1,0 +1,3 @@
+module github.com/treejamie/hackerrank-90days/testutils
+
+go 1.24.0
