+++ w1/go.mod
@@ -1,0 +1,7 @@
+module github.com/treejamie/hackerrank-90days/w1
+
+go 1.24.0
+
+require github.com/treejamie/hackerrank-90days/testutils v0.0.0
+
+replace github.com/treejamie/hackerrank-90days/testutils => ../test
